+++ frontend/src/components/Navbar.jsx
@@ -7,12 +7,13 @@ import IconButton from '@mui/material/IconButton';
 import Badge from '@mui/material/Badge';
 import MenuItem from '@mui/material/MenuItem';
 import Menu from '@mui/material/Menu';
-import AccountCircle from '@mui/icons-material/AccountCircle';
 import MailIcon from '@mui/icons-material/Mail';
 import NotificationsIcon from '@mui/icons-material/Notifications';
 import MoreIcon from '@mui/icons-material/MoreVert';
 import { Typography } from '@mui/material';
 import Profile from '../assets/photo-1633332755192-727a05c4013d.jpg'
+import {Link} from 'react-router-dom';
+import ArrowDropDownIcon from '@mui/icons-material/ArrowDropDown';
 
 
 
@@ -40,6 +41,7 @@ export default function PrimarySearchAppBar() {
     setMobileMoreAnchorEl(event.currentTarget);
   };
 
+  
   const menuId = 'primary-search-account-menu';
   const renderMenu = (
     <Menu
@@ -78,68 +80,87 @@ export default function PrimarySearchAppBar() {
       }}
       open={isMobileMenuOpen}
       onClose={handleMobileMenuClose}
+      
     >
       <MenuItem>
-        <IconButton size="large" aria-label="show 4 new mails" color="inherit">
+        <IconButton size="1rem" aria-label="show 4 new mails" color="inherit">
           <Badge variant="dot" color="error">
-            <MailIcon />
+            <MailIcon style={{fontSize:"1.5rem"}}/>
           </Badge>
         </IconButton>
-        <p>Messages</p>
+        <p style={{fontSize:"13px"}}>Messages</p>
       </MenuItem>
       <MenuItem>
         <IconButton
-          size="large"
+          size="1rem"
           aria-label="show 17 new notifications"
           color="inherit"
         >
           <Badge variant="dot" color="error">
-            <NotificationsIcon />
+            <NotificationsIcon style={{fontSize:"1.5rem"}}/>
           </Badge>
         </IconButton>
-        <p>Notifications</p>
+        <p style={{fontSize:"13px"}}>Notifications</p>
       </MenuItem>
       <MenuItem onClick={handleProfileMenuOpen}>
         <IconButton
-          size="large"
+          size="1rem"
           aria-label="account of current user"
           aria-controls="primary-search-account-menu"
           aria-haspopup="true"
           color="inherit"
         >
           {/* <AccountCircle /> */}
-          <img src={Profile} alt="Profile" width="40px" height="40px"  style={{borderRadius:"50px"}}/>
+          <img src={Profile} alt="Profile" width="35px" height="35px"  style={{borderRadius:"50px", marginLeft:"-0.4rem"}}/>
         </IconButton>
-        <p>Profile</p>
+        <p style={{fontSize:"13px"}}>Profile</p>
       </MenuItem>
     </Menu>
   );
 
+
   return (
     <Box sx={{ flexGrow: 1 }}>
-      <AppBar position="static" sx={{ backgroundColor: 'white' }}>
+      <AppBar position="static" sx={{ backgroundColor: 'white', height:"60px" }}>
         <Toolbar>
           <Box sx={{ flexGrow: 1 }} />
           <Box sx={{ display: { xs: 'none', md: 'flex' } }}>
             <IconButton size="large" aria-label="show 4 new mails" color="inherit" style={{borderRadius: "50%",  width: "60px", height: "60px", marginTop:"5px"}}>
               <Badge variant="dot" color="error">
-                <MailIcon style={{ fontSize: "25px", color: "black"  }}/>
+                <MailIcon style={{ fontSize: "1.2rem", color: "black"  }}/>
               </Badge>
             </IconButton>
+            <Link to="../pages/member/Notification.jsx" style={{textDecoration:"none"}}>
+              <IconButton
+                size="large"
+                aria-label="show 17 new notifications"
+                color="inherit"
+                style={{borderRadius: "50%",  width: "60px", height: "60px", marginTop:"5px"}}
+              >
+                <Badge variant="dot" color="error" style={{marginRight:"10px"}}>
+                <NotificationsIcon style={{ fontSize: "1.2rem", color: "black" }} />
+                </Badge>
+              </IconButton>
+            </Link>
+            <Typography  variant='subtitle1' component="div" sx={{ flexGrow: 1, mt:2, mr:1, fontSize:"14px", color:"black", fontWeight:700 }}>
+                Tharindu Gunawardhana
+                <Typography variant="subtitle2"  gutterBottom sx={{ flexGrow: 1, fontSize:"12px", color:"grey", fontWeight:500, textAlign:"right" }}>
+                  Physical Member
+                </Typography>
+            </Typography>
             <IconButton
               size="large"
-              aria-label="show 17 new notifications"
+              edge="end"
+              aria-label="account of current user"
+              aria-controls={menuId}
+              aria-haspopup="true"
+       
               color="inherit"
-              style={{borderRadius: "50%",  width: "60px", height: "60px", marginTop:"5px"}}
             >
-              <Badge variant="dot" color="error" style={{marginRight:"10px"}}>
-              <NotificationsIcon style={{ fontSize: "25px", color: "black" }} />
-              </Badge>
+              {/* <AccountCircle /> */}
+              <img src={Profile} alt="Profile" width="50px" height="50px" style={{borderRadius:"50px"}}/>
             </IconButton>
-            <Typography  variant='subtitle1' component="div" sx={{ flexGrow: 1, mt:2, mr:1, fontSize:"15px", color:"black", fontWeight:700 }}>
-                Tharindu Gunawardhana
-                <Typography variant="subtitle2"  gutterBottom sx={{ flexGrow: 1, fontSize:"12px", color:"grey", fontWeight:500 }}>Physical Member</Typography>
-            </Typography>
+           
             <IconButton
               size="large"
               edge="end"
@@ -150,7 +171,7 @@ export default function PrimarySearchAppBar() {
               color="inherit"
             >
               {/* <AccountCircle /> */}
-              <img src={Profile} alt="Profile" width="50px" height="50px" style={{borderRadius:"50px"}}/>
+              <ArrowDropDownIcon style={{color:"#000000"}}/>
             </IconButton>
           </Box>
           <Box sx={{ display: { xs: 'flex', md: 'none' } }}>
@@ -162,11 +183,12 @@ export default function PrimarySearchAppBar() {
               onClick={handleMobileMenuOpen}
               color="inherit"
             >
-              <MoreIcon />
+              <MoreIcon style={{color:"black"}}/>
             </IconButton>
           </Box>
         </Toolbar>
       </AppBar>
+
       {renderMobileMenu}
       {renderMenu}
     </Box>
