+++ frontend/src/pages/member/Shakebar.jsx
@@ -1,6 +1,8 @@
 import React, {useState} from "react";
 import Box from "@mui/material/Box";
 import { Typography,  Select, MenuItem, Button, InputLabel, FormControl} from "@mui/material";
+import Sidebar from "../../components/Sidebar";
+import Navbar from "../../components/Navbar";
 
 
 import item1 from "../../assets/images (3).jpg"
@@ -23,109 +25,122 @@ const Shakebar = () => {
 
 
   return (
-    <Box sx={{}}>
-        <Typography variant="h3" style={{ fontWeight: 700, marginTop: "1rem", textAlign:"left" }}>Shakebar</Typography>
 
-        <Box sx={{display:"flex", marginTop:"1rem"}}>
-            <FormControl style={{width:"15%"}}>
-                <InputLabel id="demo-simple-select-label">All</InputLabel>
-                <Select
-                style={{height:"85%"}}
-                labelId="demo-simple-select-label"
-                id="demo-simple-select"
-                value={item}
-                label="All"
-                onChange={handleChange}
-                >
-                <MenuItem value={10}>Vitamins</MenuItem>
-                <MenuItem value={20}>Proteins</MenuItem>
-                <MenuItem value={30}>Shakes</MenuItem>
-                </Select>
-            </FormControl>
+    <Box sx={{ flex: "1", display:"flex", mb:2}}>
+      <Box>
+        <Sidebar />
+      </Box>
+      
+      <Box component="main" sx={{flex:1 }}>
+        <Box>
+          <Navbar />
         </Box>
-        
-        <Box sx={{display:"flex", width:"95%", height:"70%", backgroundColor:"#E5E8E8", padding:"0.3rem", borderRadius:"10px", marginBottom:"2rem", marginTop:"1.5rem"}}>
-            <Box sx={{display:"flex",height:"82vh", flexWrap: "wrap",overflowY:"auto", width:"100%", backgroundColor:"white", borderRadius:"10px", padding: "1rem", margin:"0.1rem"}}>
-                <Box sx={{width:"22%",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
+
+        <Box sx={{ paddingLeft:"5rem", flex:1 }}>
+
+          <Typography variant="h3" style={{ fontWeight: 700, marginTop: "1rem", textAlign:"left" }}>Shakebar</Typography>
+
+          <Box sx={{display:"flex", marginTop:"1rem"}}>
+              <FormControl style={{width:"15%"}}>
+                  <InputLabel id="demo-simple-select-label">All</InputLabel>
+                  <Select
+                  style={{height:"85%"}}
+                  labelId="demo-simple-select-label"
+                  id="demo-simple-select"
+                  value={item}
+                  label="All"
+                  onChange={handleChange}
+                  >
+                  <MenuItem value={10}>Vitamins</MenuItem>
+                  <MenuItem value={20}>Proteins</MenuItem>
+                  <MenuItem value={30}>Shakes</MenuItem>
+                  </Select>
+              </FormControl>
+          </Box>
+
+          <Box sx={{display:"flex", width:"96%", height:"70%", backgroundColor:"#E5E8E8", padding:"0.3rem", borderRadius:"10px", marginBottom:"2rem", marginTop:"1.5rem"}}>
+              <Box sx={{display:"flex",height:"82vh", flexWrap: "wrap",overflowY:"auto", width:"100%", backgroundColor:"white", borderRadius:"10px", padding: "1rem", margin:"0.1rem"}}>
+                  
+              <Box sx={{width:"22%",textAlign:"center",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
                     <img src={item1} alt="item" style={{width:"85%", height:"65%"}}></img>
                     <Typography variant="h6" style={{fontWeight: 700}}>Vitamin C</Typography>
                     <Typography variant="h6" style={{fontWeight: 500}}>Rs: 2300</Typography>
                     <Button variant="contained" style={{backgroundColor:"#FFBF00", color:"black", fontWeight:"700", width: "50%"}}>Buy</Button>
                 </Box>
-                <Box sx={{width:"22%",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
+                <Box sx={{width:"22%",textAlign:"center",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
                     <img src={item2} alt="item"  style={{width:"85%", height:"65%"}}></img>
                     <Typography variant="h6" style={{fontWeight: 700}}>Vitamin C</Typography>
                     <Typography variant="h6" style={{fontWeight: 500}}>Rs: 2300</Typography>
                     <Button variant="contained" style={{backgroundColor:"#FFBF00", color:"black", fontWeight:"700", width: "50%"}}>Buy</Button>
                 </Box>
-                <Box sx={{width:"22%",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
+                <Box sx={{width:"22%",textAlign:"center",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
                     <img src={item3} alt="item"  style={{width:"85%", height:"65%"}}></img>
                     <Typography variant="h6" style={{fontWeight: 700}}>Vitamin C</Typography>
                     <Typography variant="h6" style={{fontWeight: 500}}>Rs: 2300</Typography>
                     <Button variant="contained" style={{backgroundColor:"#FFBF00", color:"black", fontWeight:"700", width: "50%"}}>Buy</Button>
                 </Box>
-                <Box sx={{width:"22%",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
+                <Box sx={{width:"22%",textAlign:"center",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
                     <img src={item4} alt="item"  style={{width:"60%", height:"65%"}}></img>
                     <Typography variant="h6" style={{fontWeight: 700}}>Vitamin C</Typography>
                     <Typography variant="h6" style={{fontWeight: 500}}>Rs: 2300</Typography>
                     <Button variant="contained" style={{backgroundColor:"#FFBF00", color:"black", fontWeight:"700", width: "50%"}}>Buy</Button>
                 </Box>
-                <Box sx={{width:"22%",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
+                <Box sx={{width:"22%",textAlign:"center",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
                     <img src={item5} alt="item"  style={{width:"85%", height:"65%"}}></img>
                     <Typography variant="h6" style={{fontWeight: 700}}>Vitamin C</Typography>
                     <Typography variant="h6" style={{fontWeight: 500}}>Rs: 2300</Typography>
                     <Button variant="contained" style={{backgroundColor:"#FFBF00", color:"black", fontWeight:"700", width: "50%"}}>Buy</Button>
                 </Box>
-                <Box sx={{width:"22%",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
+                <Box sx={{width:"22%",textAlign:"center",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
                     <img src={item6} alt="item"  style={{width:"85%", height:"65%"}}></img>
                     <Typography variant="h6" style={{fontWeight: 700}}>Vitamin C</Typography>
                     <Typography variant="h6" style={{fontWeight: 500}}>Rs: 2300</Typography>
                     <Button variant="contained" style={{backgroundColor:"#FFBF00", color:"black", fontWeight:"700", width: "50%"}}>Buy</Button>
                 </Box>
-                <Box sx={{width:"22%",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
+                <Box sx={{width:"22%",textAlign:"center",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
                     <img src={item7} alt="item"  style={{width:"60%", height:"65%"}}></img>
                     <Typography variant="h6" style={{fontWeight: 700}}>Vitamin C</Typography>
                     <Typography variant="h6" style={{fontWeight: 500}}>Rs: 2300</Typography>
                     <Button variant="contained" style={{backgroundColor:"#FFBF00", color:"black", fontWeight:"700", width: "50%"}}>Buy</Button>
                 </Box>
-                <Box sx={{width:"22%",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
+                <Box sx={{width:"22%",textAlign:"center",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
                     <img src={item8} alt="item"  style={{width:"60%", height:"65%"}}></img>
                     <Typography variant="h6" style={{fontWeight: 700}}>Vitamin C</Typography>
                     <Typography variant="h6" style={{fontWeight: 500}}>Rs: 2300</Typography>
                     <Button variant="contained" style={{backgroundColor:"#FFBF00", color:"black", fontWeight:"700", width: "50%"}}>Buy</Button>
                 </Box>
-                <Box sx={{width:"22%",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
+                <Box sx={{width:"22%",textAlign:"center",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
                     <img src={item1} alt="item"  style={{width:"85%", height:"65%"}}></img>
                     <Typography variant="h6" style={{fontWeight: 700}}>Vitamin C</Typography>
                     <Typography variant="h6" style={{fontWeight: 500}}>Rs: 2300</Typography>
                     <Button variant="contained" style={{backgroundColor:"#FFBF00", color:"black", fontWeight:"700", width: "50%"}}>Buy</Button>
                 </Box>
-                <Box sx={{width:"22%",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
+                <Box sx={{width:"22%",textAlign:"center",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
                     <img src={item1} alt="item"  style={{width:"85%", height:"65%"}}></img>
                     <Typography variant="h6" style={{fontWeight: 700}}>Vitamin C</Typography>
                     <Typography variant="h6" style={{fontWeight: 500}}>Rs: 2300</Typography>
                     <Button variant="contained" style={{backgroundColor:"#FFBF00", color:"black", fontWeight:"700", width: "50%"}}>Buy</Button>
                 </Box>
                
-                <Box sx={{width:"22%",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
+                <Box sx={{width:"22%",textAlign:"center",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
                     <img src={item1} alt="item"  style={{width:"85%", height:"65%"}}></img>
                     <Typography variant="h6" style={{fontWeight: 700}}>Vitamin C</Typography>
                     <Typography variant="h6" style={{fontWeight: 500}}>Rs: 2300</Typography>
                     <Button variant="contained" style={{backgroundColor:"#FFBF00", color:"black", fontWeight:"700", width: "50%"}}>Buy</Button>
                 </Box>
-                <Box sx={{width:"22%",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
+                <Box sx={{width:"22%",textAlign:"center",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
                     <img src={item1} alt="item"  style={{width:"85%", height:"65%"}}></img>
                     <Typography variant="h6" style={{fontWeight: 700}}>Vitamin C</Typography>
                     <Typography variant="h6" style={{fontWeight: 500}}>Rs: 2300</Typography>
                     <Button variant="contained" style={{backgroundColor:"#FFBF00", color:"black", fontWeight:"700", width: "50%"}}>Buy</Button>
                 </Box>
-                <Box sx={{width:"22%",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
+                <Box sx={{width:"22%",textAlign:"center",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
                     <img src={item1} alt="item"  style={{width:"85%", height:"65%"}}></img>
                     <Typography variant="h6" style={{fontWeight: 700}}>Vitamin C</Typography>
                     <Typography variant="h6" style={{fontWeight: 500}}>Rs: 2300</Typography>
                     <Button variant="contained" style={{backgroundColor:"#FFBF00", color:"black", fontWeight:"700", width: "50%"}}>Buy</Button>
                 </Box>
-                <Box sx={{width:"22%",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
+                <Box sx={{width:"22%",textAlign:"center",justifyContent: "center", alignItems:"center",height:"60%", cursor:"pointer", border:"2px solid white", borderRadius:"10px" , padding:"1rem", marginRight:"2rem", marginBottom:"1rem", boxShadow: 'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px', '&:hover': {borderColor: '#FFBF00',  transition: "ease 0.5s"}}}>
                     <img src={item1} alt="item"  style={{width:"85%", height:"65%"}}></img>
                     <Typography variant="h6" style={{fontWeight: 700}}>Vitamin C</Typography>
                     <Typography variant="h6" style={{fontWeight: 500}}>Rs: 2300</Typography>
@@ -133,13 +148,21 @@ const Shakebar = () => {
                 </Box>
 
 
-                
-            </Box>
-           
+                  
+              </Box>
+            
+          </Box>
+
+
         </Box>
-
+          
       
+     
+      </Box>
+     
     </Box>
+
+    
   );
 };
